Ready for first release
Player can be started from editor and load an image as source.
To use a video as source still a virtual camera is needed.
The player will load a camera if starded without editor so you can use
it on an mobile device if the needed files are at the right place
described in readme